Add filebeat role to ship logs to logstash
Beaver is no longer maintained (or just barely maintained) so to avoid
depending on software that we cannot receive fixes for, we are
preemptively switching to Filebeat which was created by Elastic.co.

Filebeat provides "lightweight" log shipping to logstash with some
tweaks to logstash to parse the format that Filebeat uses and to open
a port specifically for Filebeat to use.

To enable Filebeat, we have a minimal viable configuration file that is
templated out by the role. Given the absence of the config_template
module used by OpenStack-Ansible, we use a more clever way of templating
out the YAML necessary to configure Filebeat.

Further, since we no longer need beaver (and because deploying it can
make testing Filebeat a little confusing) we also remove the beaver role
and replace it in our setup-logging playbook with our new filebeat
playbook.

Connects rcbops/u-suk-dev#177