Remove tag added to log events by logstash plugin
The logstash-input-beats plugin is necessary to use filebeat to ship
logs to logstash. However, it defaults to always adding a tag to each
log event that indicates what codec it used to process the beats input.

Until https://github.com/logstash-plugins/logstash-input-beats/pull/114
is merged (in one shape or another) we need to just remove the tag from
all of our events.

Connects rcbops/u-suk-dev#261